requirements.txt
Flask
requests
python-dotenv
pyyaml
pydantic